LayoutTests/platform/blackberry/fast/dom/attribute-namespaces-expected.txt

CONSOLE MESSAGE: line 7: TypeError: Result of expression 't' [null] is not an object.
This tests that attributes will have the correct namespace, prefix and local name set.
